packages/python-ml-service/requirements.txt
grpcio>=1.58.0
grpcio-health-checking>=1.58.0
grpcio-reflection>=1.58.0
grpcio-tools>=1.58.0
pdfminer.six>=20221105
prometheus-client>=0.17.1
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
structlog>=23.1.0
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
PyPDF2>=3.0.1
protobuf>=4.24.3
tenacity>=8.2.3
numpy>=1.26.0
scikit-learn>=1.3.1
pandas>=2.1.1
crawl4ai>=0.4.248